web/requirements.txt
Django==1.3.1
-e git://github.com/chrisgilmerproj/django-gravatar.git#egg=gravatar
django-piston
